move render_strict helper to lib module
